requirements.txt
addict==2.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bilby<=2.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
cycler==0.11.0
Cython==0.29.28
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
emcee==3.1.4
fonttools==4.37.1
future==0.18.2
idna==3.3
imageio==2.19.1
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==7.33.0
jedi==0.18.1
jinja2<3.1.0`
jupyter-client==7.0.6
jupyter-core==4.9.2
kiwisolver==1.4.4
Markdown==3.3.6
matplotlib==3.5.3
matplotlib-inline==0.1.3
model-index==0.1.11
mkdocs-jupyter
mkdocstrings
mkdocs-toc-md
mkdocstrings-python
myst-parser
nest-asyncio==1.5.5
networkx==2.6.3
numpy==1.21.6
opencv-python==4.5.5.64
openmim==0.1.5
ordered-set==4.1.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==21.2.2
prompt-toolkit==3.0.29
psutil==5.8.0
ptyprocess==0.7.0
pycocotools==2.0.6
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==19.0.2
requests==2.28.2
scikit-image==0.19.2
scipy==1.7.3
setuptools==60.2.0
six==1.16.0
tabulate==0.8.9
terminaltables==3.1.10
tifffile==2021.11.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.3.0
urllib3==1.26.9
wcwidth==0.2.5
wheel==0.37.1
yapf==0.32.0
zipp==3.8.0
